feature/music : 코드 수정

+++ src/main/java/org/scoula/mapper/MusicMapper.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 public interface MusicMapper {
     public List<MusicVO> getList();  // 전체 목록 조회
-    public MusicVO get(int id);  // 개별 목록 조회
+    public MusicVO get(Long id);  // 개별 목록 조회
     public void create(MusicVO music);  // 등록
     public int update(MusicVO music);  // 수정
-    public int delete(int id);  // 삭제
+    public int delete(Long id);  // 삭제
 }
